refactor(quick-panel): rename forked panelEngine files to naming spec
Addresses @kangfenmao's review: new files landing on main must follow
docs/references/naming-conventions.md, so the forked panelEngine no longer
ships generic lowercase names:

- hook.ts -> useQuickPanel.ts (§3.2 hook-primary file)
- provider.tsx -> QuickPanelProvider.tsx
- view.tsx -> QuickPanelView.tsx
- list.tsx -> QuickPanelList.tsx (per its list-primitives responsibility)

Test files + the barrel re-exports follow; external consumers import via the
unchanged barrel, so nothing else moves. No mocks reference these paths.
