+++ src/renderer/components/chat/composer/panelEngine/QuickPanelList.tsx
