+++ src/renderer/components/chat/composer/panelEngine/QuickPanelProvider.tsx
