+++ src/renderer/components/chat/composer/panelEngine/index.ts
@@ -1,19 +1,19 @@
 /**
  * `panelEngine` is a temporary fork of `components/QuickPanel`, created so the v2
  * composer can evolve its quick-panel (inputAdapter / readOnly / keydown-dispatch
- * model in `provider`, `view`, `types`) without touching the live v1 QuickPanel.
+ * model in `QuickPanelProvider`, `QuickPanelView`, `types`) without touching the live v1 QuickPanel.
  *
- * - `provider` / `view` / `types` intentionally diverge from v1 — fix here, not there.
+ * - `QuickPanelProvider` / `QuickPanelView` / `types` intentionally diverge from v1 — fix here, not there.
  * - `defaultStrategies` is the ONLY byte-identical copy of v1 — keep the two in sync until
- *   collapse. `hook` diverges (adds `useOptionalQuickPanel`); `list` / `heights` were newly
- *   extracted from the forked `view` and have no v1 twin.
+ *   collapse. `useQuickPanel` diverges (adds `useOptionalQuickPanel`); `QuickPanelList` / `heights`
+ *   were newly extracted from the forked `QuickPanelView` and have no v1 twin.
  *
  * Collapse plan: at the pages switchover (when the v1 Inputbar is deleted) this fork
  * merges back into `components/QuickPanel` and this directory is removed.
  */
 export * from './defaultStrategies'
-export * from './hook'
-export * from './list'
-export * from './provider'
+export * from './QuickPanelList'
+export * from './QuickPanelProvider'
+export * from './QuickPanelView'
 export * from './types'
-export * from './view'
+export * from './useQuickPanel'
